Version cnage, update readme, add nuget double->float

+++ Perfon.Core/Notifications/IThresholdNotification.cs
@@ -13,7 +13,7 @@ namespace Perfon.Core.Notifications
     /// </summary>
     public interface IThresholdNotification
     {
-        double ThresholdValue { get; }
+        float ThresholdValue { get; }
 
         bool IsThresholdViolated {get;}
